pyproject.toml
[tool.poetry]
name = "blog"
version = "0.1.0"
description = ""
authors = ["Wasil P <it.dzemidok@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
psycopg2-binary = "^2.9.7"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
pillow = "^10.0.0"
django-crispy-forms = "^2.0"
crispy-bootstrap5 = "^0.7"
celery = "^5.3.4"
aiogram = "^3.0.0"
requests = "^2.31.0"
redis = "^5.0.0"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
